+++ cmd/controlplane/management_controller.go
@@ -16,7 +16,7 @@ import (
 	"github.com/akuity/kargo/internal/api/kubernetes"
 	"github.com/akuity/kargo/internal/controller/management/namespaces"
 	"github.com/akuity/kargo/internal/controller/management/projects"
-	"github.com/akuity/kargo/internal/controller/management/serviceaccounts"
+	"github.com/akuity/kargo/internal/controller/management/upgrade"
 	"github.com/akuity/kargo/internal/logging"
 	"github.com/akuity/kargo/internal/os"
 	versionpkg "github.com/akuity/kargo/internal/version"
@@ -79,7 +79,7 @@ func (o *managementControllerOptions) run(ctx context.Context) error {
 		return fmt.Errorf("error setting up Projects reconciler: %w", err)
 	}
 
-	if err := serviceaccounts.SetupServiceAccountReconcilerWithManager(kargoMgr); err != nil {
+	if err := upgrade.SetupServiceAccountReconcilerWithManager(kargoMgr); err != nil {
 		return fmt.Errorf("error setting up ServiceAccount reconciler: %w", err)
 	}
