--- internal/controller/management/serviceaccounts/serviceaccounts_test.go
@@ -1,1 +1,0 @@
-package serviceaccounts
