+++ internal/controller/management/upgrade/serviceaccounts.go
@@ -1,11 +1,11 @@
-package serviceaccounts
+package upgrade
 
 import (
 	"context"
+	"fmt"
 
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/event"
@@ -25,12 +25,12 @@ const (
 // ServiceAccountReconciler Reconciles for ServiceAccounts that need annotation migration
 type ServiceAccountReconciler struct {
 	client.Client
-	Scheme *runtime.Scheme
-	Logger *logging.Logger
 }
 
-// SetupWithManager sets up the reconciler with the manager
-func (r *ServiceAccountReconciler) SetupWithManager(mgr ctrl.Manager) error {
+// SetupServiceAccountReconcilerWithManager initializes a reconciler
+// that upgrades annotations on ServiceAccount resources for v0.9.x
+// compatibility and registers that reconciler with the provided Manager.
+func SetupServiceAccountReconcilerWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&corev1.ServiceAccount{}).
 		WithEventFilter(predicate.Funcs{
@@ -40,55 +40,50 @@ func (r *ServiceAccountReconciler) SetupWithManager(mgr ctrl.Manager) error {
 			CreateFunc: func(e event.CreateEvent) bool {
 				return hasOldAnnotations(e.Object.GetAnnotations())
 			},
-			GenericFunc: func(e event.GenericEvent) bool {
-				return hasOldAnnotations(e.Object.GetAnnotations())
+			DeleteFunc: func(event.DeleteEvent) bool {
+				return false
+			},
+			GenericFunc: func(event.GenericEvent) bool {
+				return false
 			},
 		}).
-		Complete(r)
+		Complete(&ServiceAccountReconciler{
+			Client: mgr.GetClient(),
+		})
 }
 
 // Reconcile handles the reconciliation logic for migrating annotations
 func (r *ServiceAccountReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	sa := &corev1.ServiceAccount{}
+	logger := logging.LoggerFromContext(ctx)
 	err := r.Get(ctx, req.NamespacedName, sa)
 	if err != nil {
 		if errors.IsNotFound(err) {
 			return ctrl.Result{}, nil
 		}
-		r.Logger.Error(err, "Failed to get ServiceAccount")
-		return ctrl.Result{}, err
+		return ctrl.Result{}, fmt.Errorf("failed to get ServiceAccount: %w", err)
 	}
 
-	annotations := sa.GetAnnotations()
-	if annotations == nil {
-		return ctrl.Result{}, nil
-	}
+	annotations := sa.Annotations
 
-	updated := false
 	if val, ok := annotations[OldAnnotationKeySub]; ok {
 		annotations[rbacapi.AnnotationKeyOIDCClaim("sub")] = val
 		delete(annotations, OldAnnotationKeySub)
-		updated = true
 	}
 	if val, ok := annotations[OldAnnotationKeyEmail]; ok {
 		annotations[rbacapi.AnnotationKeyOIDCClaim("email")] = val
 		delete(annotations, OldAnnotationKeyEmail)
-		updated = true
 	}
 	if val, ok := annotations[OldAnnotationKeyGroups]; ok {
 		annotations[rbacapi.AnnotationKeyOIDCClaim("groups")] = val
 		delete(annotations, OldAnnotationKeyGroups)
-		updated = true
 	}
 
-	if updated {
-		sa.SetAnnotations(annotations)
-		if err := r.Update(ctx, sa); err != nil {
-			r.Logger.Error(err, "Failed to update ServiceAccount annotations")
-			return ctrl.Result{}, err
-		}
-		r.Logger.Info("Successfully migrated ServiceAccount annotations", "ServiceAccount", req.NamespacedName)
+	sa.SetAnnotations(annotations)
+	if err := r.Update(ctx, sa); err != nil {
+		return ctrl.Result{}, fmt.Errorf("failed to update ServiceAccount annotations: %w", err)
 	}
+	logger.Info("Successfully migrated ServiceAccount annotations", "ServiceAccount", req.NamespacedName)
 
 	return ctrl.Result{}, nil
 }
@@ -102,12 +97,4 @@ func hasOldAnnotations(annotations map[string]string) bool {
 	_, hasEmail := annotations[OldAnnotationKeyEmail]
 	_, hasGroups := annotations[OldAnnotationKeyGroups]
 	return hasSub || hasEmail || hasGroups
-}
-
-func SetupServiceAccountReconcilerWithManager(mgr ctrl.Manager) error {
-	return (&ServiceAccountReconciler{
-		Client: mgr.GetClient(),
-		Scheme: mgr.GetScheme(),
-		Logger: logging.NewLogger(logging.InfoLevel),
-	}).SetupWithManager(mgr)
 }
